# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ninja Ads Earn API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  balance: number;
  totalEarned: number;
  adsWatched: number;
  createdAt: string;
}

export interface UserStats {
  balance: number;
  totalEarned: number;
  adsWatched: number;
  todayEarned: number;
  todayAds: number;
  pendingWithdrawals: number;
  referralCount: number;
}

export interface AdCompleteInput {
  telegramId: string;
  username?: string;
  firstName?: string;
}

export interface AdReward {
  earned: number;
  newBalance: number;
  totalEarned: number;
}

export interface AdView {
  id: number;
  earned: number;
  createdAt: string;
}

export interface WithdrawalInput {
  telegramId: string;
  amount: number;
  method: string;
  destination: string;
}

export interface Withdrawal {
  id: number;
  amount: number;
  method: string;
  destination: string;
  status: string;
  createdAt: string;
}

export interface LeaderboardEntry {
  rank: number;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  totalEarned: number;
  adsWatched: number;
}

export interface PlatformSettings {
  earnPerAd: number;
  referralReward: number;
  minWithdrawal: number;
  /** @nullable */
  addTaskUrl?: string | null;
  dailyWatchLimit?: number;
  withdrawalAdsRequired?: number;
  airdropEnded?: boolean;
  /** @nullable */
  airdropChannel?: string | null;
  /** @nullable */
  monetagScript?: string | null;
  /** @nullable */
  extraAdminIds?: string | null;
}

export interface AdminSettingsInput {
  adminTelegramId: string;
  earnPerAd?: number;
  referralReward?: number;
  minWithdrawal?: number;
  addTaskUrl?: string;
  dailyWatchLimit?: number;
  withdrawalAdsRequired?: number;
  airdropEnded?: boolean;
  airdropChannel?: string;
  monetagScript?: string;
  extraAdminIds?: string;
}

export interface Task {
  id: number;
  title: string;
  description: string;
  reward: number;
  /** @nullable */
  url?: string | null;
  isActive: boolean;
  createdAt: string;
}

export type TaskWithStatusSubmissionStatus =
  (typeof TaskWithStatusSubmissionStatus)[keyof typeof TaskWithStatusSubmissionStatus];

export const TaskWithStatusSubmissionStatus = {
  none: "none",
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface TaskWithStatus {
  id: number;
  title: string;
  description: string;
  reward: number;
  /** @nullable */
  url?: string | null;
  isActive: boolean;
  completed: boolean;
  submissionStatus: TaskWithStatusSubmissionStatus;
  createdAt: string;
}

export interface TaskCompleteInput {
  telegramId: string;
}

export interface TaskReward {
  earned: number;
  newBalance: number;
}

export interface AdminTaskInput {
  adminTelegramId: string;
  title: string;
  description: string;
  reward: number;
  url?: string;
  isActive?: boolean;
}

export interface AdminAuthInput {
  adminTelegramId: string;
}

export interface ReferralItem {
  /** @nullable */
  username: string | null;
  /** @nullable */
  firstName?: string | null;
  joinedAt: string;
  rewardGiven: boolean;
}

export interface ReferralStats {
  referralCode: string;
  totalReferrals: number;
  totalEarned: number;
  referrals: ReferralItem[];
}

export interface AdminWithdrawalUser {
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
}

export interface AdminWithdrawal {
  id: number;
  amount: number;
  method: string;
  destination: string;
  status: string;
  createdAt: string;
  user: AdminWithdrawalUser;
}

export interface AdminStats {
  totalUsers: number;
  totalAdsWatched: number;
  totalEarned: number;
  totalWithdrawn: number;
  pendingWithdrawals: number;
  activeTasks: number;
}

export interface OkResponse {
  ok: boolean;
}

export type GetMeParams = {
  telegramId: string;
  username?: string;
  firstName?: string;
  referralCode?: string;
};

export type GetMyStatsParams = {
  telegramId: string;
};

export type GetAdHistoryParams = {
  telegramId: string;
  limit?: number;
};

export type GetWithdrawalsParams = {
  telegramId: string;
};

export type GetLeaderboardParams = {
  limit?: number;
};

export type GetTasksParams = {
  telegramId: string;
};

export type GetReferralStatsParams = {
  telegramId: string;
};

export type GetAdminWithdrawalsParams = {
  adminTelegramId: string;
  status?: string;
};

export type GetAdminStatsParams = {
  adminTelegramId: string;
};

export type GetAdminUsersParams = {
  adminTelegramId: string;
  limit?: number;
};