# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ninja Ads Earn API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ads/complete:
    post:
      operationId: completeAd
      tags: [ads]
      summary: Record completed ad view and credit user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdCompleteInput"
      responses:
        "200":
          description: Reward credited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdReward"
  /ads/history:
    get:
      operationId: getAdHistory
      tags: [ads]
      summary: Get user ad watch history
      parameters:
        - name: telegramId
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Ad history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdView"
...
  /admin/settings:
    put:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update platform settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
...
  /admin/withdrawals:
    get:
      operationId: getAdminWithdrawals
      tags: [admin]
      summary: Get all pending withdrawals (admin only)
      parameters:
        - name: adminTelegramId
          in: query
          required: true
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            default: pending
      responses:
        "200":
          description: All withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminWithdrawal"
  /admin/withdrawals/{withdrawalId}/approve:
    post:
      operationId: approveWithdrawal
      tags: [admin]
      summary: Approve a withdrawal (admin only)
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAuthInput"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /admin/withdrawals/{withdrawalId}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [admin]
      summary: Reject a withdrawal (admin only)
      parameters:
        - name: withdrawalId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAuthInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get platform stats (admin only)
      parameters:
        - name: adminTelegramId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users (admin only)
      parameters:
        - name: adminTelegramId
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
components:
  schemas:
...
    User:
      type: object
      required: [id, telegramId, balance, totalEarned, adsWatched, createdAt]
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        balance:
          type: number
        totalEarned:
          type: number
        adsWatched:
          type: integer
        createdAt:
          type: string
...
    AdCompleteInput:
      type: object
      required: [telegramId]
      properties:
        telegramId:
          type: string
        username:
          type: string
        firstName:
          type: string
    AdReward:
      type: object
      required: [earned, newBalance, totalEarned]
      properties:
        earned:
          type: number
        newBalance:
          type: number
        totalEarned:
          type: number
    AdView:
      type: object
      required: [id, earned, createdAt]
      properties:
        id:
          type: integer
        earned:
          type: number
        createdAt:
          type: string
...
    PlatformSettings:
      type: object
      required: [earnPerAd, referralReward, minWithdrawal]
      properties:
        earnPerAd:
          type: number
        referralReward:
          type: number
        minWithdrawal:
          type: number
        addTaskUrl:
          type: ["string", "null"]
    AdminSettingsInput:
      type: object
      required: [adminTelegramId]
      properties:
        adminTelegramId:
          type: string
        earnPerAd:
          type: number
        referralReward:
          type: number
        minWithdrawal:
          type: number
        addTaskUrl:
          type: string
...
    AdminAuthInput:
      type: object
      required: [adminTelegramId]
      properties:
        adminTelegramId:
          type: string
...
    AdminWithdrawal:
      type: object
      required: [id, amount, method, destination, status, createdAt, user]
      properties:
        id:
          type: integer
        amount:
          type: number
        method:
          type: string
        destination:
          type: string
        status:
          type: string
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/AdminWithdrawalUser"
    AdminWithdrawalUser:
      type: object
      required: [telegramId]
      properties:
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
    AdminStats:
      type: object
      required: [totalUsers, totalAdsWatched, totalEarned, totalWithdrawn, pendingWithdrawals, activeTasks]
      properties:
        totalUsers:
          type: integer
        totalAdsWatched:
          type: integer
        totalEarned:
          type: number
        totalWithdrawn:
          type: number
        pendingWithdrawals:
          type: number
        activeTasks:
          type: integer
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean