# lib/api-zod/src/generated/types/adminWithdrawal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ninja Ads Earn API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminWithdrawalUser } from "./adminWithdrawalUser";

export interface AdminWithdrawal {
  id: number;
  amount: number;
  method: string;
  destination: string;
  status: string;
  createdAt: string;
  user: AdminWithdrawalUser;
}