# lib/api-zod/src/generated/types/adminWithdrawalUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ninja Ads Earn API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminWithdrawalUser {
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
}