# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ninja Ads Earn API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  balance: number;
  totalEarned: number;
  adsWatched: number;
  createdAt: string;
}